refactor: Split price hooks
