+++ src/state/farms/hooks.ts
@@ -90,3 +90,15 @@ export const useLpTokenPrice = (symbol: string) => {
 
   return lpTokenPrice
 }
+
+// /!\ Deprecated , use the BUSD hook in /hooks
+
+export const usePriceBnbBusd = (): BigNumber => {
+  const bnbBusdFarm = useFarmFromPid(252)
+  return new BigNumber(bnbBusdFarm.quoteToken.busdPrice)
+}
+
+export const usePriceCakeBusd = (): BigNumber => {
+  const cakeBnbFarm = useFarmFromPid(251)
+  return new BigNumber(cakeBnbFarm.token.busdPrice)
+}
