+++ src/views/Farms/Farms.tsx
@@ -7,8 +7,7 @@ import { ChainId } from '@pancakeswap/sdk'
 import styled from 'styled-components'
 import FlexLayout from 'components/Layout/Flex'
 import Page from 'components/Layout/Page'
-import { usePriceCakeBusd } from 'state/hooks'
-import { useFarms, usePollFarmsData } from 'state/farms/hooks'
+import { useFarms, usePollFarmsData, usePriceCakeBusd } from 'state/farms/hooks'
 import usePersistState from 'hooks/usePersistState'
 import { Farm } from 'state/types'
 import { useTranslation } from 'contexts/Localization'
