+++ src/views/Home/components/PredictionPromotionCard.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Card, CardBody, ArrowForwardIcon, Link, Button, Text, Skeleton, Flex } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { getTotalWon } from 'state/predictions/helpers'
-import { usePriceBnbBusd } from 'state/hooks'
+import { usePriceBnbBusd } from 'state/farms/hooks'
 import useRefresh from 'hooks/useRefresh'
 import useIntersectionObserver from 'hooks/useIntersectionObserver'
 import Balance from '../../../components/Balance'
