+++ src/views/Predictions/components/History/BetResult.tsx
@@ -3,7 +3,8 @@ import { useWeb3React } from '@web3-react/core'
 import { Box, Flex, Heading, Text, PrizeIcon, BlockIcon, LinkExternal, useTooltip, InfoIcon } from '@pancakeswap/uikit'
 import { useAppDispatch } from 'state'
 import { useTranslation } from 'contexts/Localization'
-import { useGetRewardRate, usePriceBnbBusd } from 'state/hooks'
+import { useGetRewardRate } from 'state/hooks'
+import { usePriceBnbBusd } from 'state/farms/hooks'
 import styled from 'styled-components'
 import { Bet, BetPosition } from 'state/types'
 import { fetchLedgerData, markBetHistoryAsCollected } from 'state/predictions'
