+++ src/views/Predictions/components/History/PnlTab/PnlTab.tsx
@@ -6,7 +6,8 @@ import { useTranslation } from 'contexts/Localization'
 import { getRoundResult, Result } from 'state/predictions/helpers'
 import { getBscScanLink } from 'utils'
 import store from 'state'
-import { useGetCurrentEpoch, usePriceBnbBusd } from 'state/hooks'
+import { useGetCurrentEpoch } from 'state/hooks'
+import { usePriceBnbBusd } from 'state/farms/hooks'
 import { Bet, BetPosition } from 'state/types'
 import { formatBnb, getMultiplier, getNetPayout } from '../helpers'
 import PnlChart from './PnlChart'
